pull from master

+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: yquaro <yquaro@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/22 12:05:21 by yquaro            #+#    #+#             */
-/*   Updated: 2019/05/20 17:27:08 by yquaro           ###   ########.fr       */
+/*   Updated: 2019/05/05 19:02:16 by yquaro           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -85,7 +85,7 @@ void				ft_lstadd(t_list **alst, t_list *new);
 void				ft_lstiter(t_list *lst, void (*f)(t_list *elem));
 t_list				*ft_lstmap(t_list *lst, t_list *(*f)(t_list *elem));
 int					ft_height(char *s, char c);
-int					ft_abs(int num);
+int					ft_abs(long long int num);
 int					ft_pow(int num, int power);
 int					ft_isspace(int c);
 int					ft_numblen(int n);
@@ -97,8 +97,7 @@ char				*ft_strencut(char *str, int n, char *choice);
 char				*ft_addsymb(char *str, char c, int num);
 char				*ft_addsymbend(char *str, char c, int num);
 char				*ft_strglue(char *str, const char *left, const char *right);
-char				*ft_strnglue(char *str, const char *left, const char *right,
-						size_t n);
+char				*ft_strnglue(char *str, const char *left, const char *right, size_t n);
 char				*ft_stradd(char *str, char c);
 char				*ft_strnadd(char *str, char c, size_t n);
